Fix x64 compilation error in r4673. Review URL: http://codereview.chromium.org/2129010
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@4674 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/profile-generator.cc
@@ -545,7 +545,7 @@ CpuProfile* CpuProfilesCollection::GetProfile(int security_token_id,
                                                 false);
   int index;
   if (entry != NULL) {
-    index = reinterpret_cast<int>(entry->value);
+    index = static_cast<int>(reinterpret_cast<intptr_t>(entry->value));
   } else {
     return NULL;
   }
